+++ parser_library/src/diagnostic_op.h
@@ -313,6 +313,22 @@ struct diagnostic_op
 
     static diagnostic_op error_A165_POP_USING(const range& range);
 
+    static diagnostic_op error_A166_GOFF_required(const range& range);
+
+    static diagnostic_op error_A167_CATTR_label(const range& range);
+
+    static diagnostic_op error_A168_XATTR_label(const range& range);
+
+    static diagnostic_op error_A169_no_section(const range& range);
+
+    static diagnostic_op error_A170_section_type_mismatch(const range& range);
+
+    static diagnostic_op warn_A171_operands_ignored(const range& range);
+
+    static diagnostic_op warn_A172_psect_redefinition(const range& range);
+
+    static diagnostic_op error_A173_invalid_psect(const range& range);
+
     // operand parameters
 
     static diagnostic_op error_A200_SCOPE_param(std::string_view instr_name, const range& range);