+++ parser_library/src/processing/instruction_sets/asm_processor.cpp
@@ -835,6 +835,8 @@ asm_processor::process_table_t asm_processor::create_table()
     table.emplace(context::id_index("CXD"), [this](rebuilt_statement&& stmt) { process_CXD(std::move(stmt)); });
     table.emplace(context::id_index("TITLE"), [this](rebuilt_statement&& stmt) { process_TITLE(std::move(stmt)); });
     table.emplace(context::id_index("PUNCH"), [this](rebuilt_statement&& stmt) { process_PUNCH(std::move(stmt)); });
+    table.emplace(context::id_index("CATTR"), [this](rebuilt_statement&& stmt) { process_CATTR(std::move(stmt)); });
+    table.emplace(context::id_index("XATTR"), [this](rebuilt_statement&& stmt) { process_XATTR(std::move(stmt)); });
 
     return table;
 }
@@ -1462,6 +1464,158 @@ void asm_processor::process_PUNCH(rebuilt_statement&& stmt)
     utils::append_utf8_sanitized(sanitized, text);
 
     output->punch(sanitized);
+}
+
+asm_processor::cattr_ops_result asm_processor::cattr_ops(const semantics::operands_si& ops)
+{
+    using enum semantics::operand_type;
+    if (ops.value.empty())
+        return {};
+    if (ops.value.size() == 2 && ops.value.front()->type == EMPTY && ops.value.back()->type == EMPTY)
+        return {};
+
+    for (const auto& op : ops.value)
+    {
+        const auto* asm_op = op->access_asm();
+        if (!asm_op)
+            continue;
+        const auto* complex = asm_op->access_complex();
+        if (!complex)
+            continue;
+        if (complex->value.identifier != "PART" || complex->value.values.size() != 1)
+            continue;
+        const auto* str = dynamic_cast<const semantics::complex_assembler_operand::string_value_t*>(
+            complex->value.values.front().get());
+        auto [_, result] = hlasm_ctx.try_get_symbol_name(str->value);
+
+        return { ops.value.size() - 1, result, complex->operand_range };
+    }
+
+    return { ops.value.size(), {}, {} };
+}
+
+void asm_processor::handle_cattr_ops(context::id_index class_name,
+    context::id_index part_name,
+    const range& part_rng,
+    size_t op_count,
+    const rebuilt_statement& stmt)
+{
+    assert(!class_name.empty());
+    auto* class_name_sect = hlasm_ctx.ord_ctx.get_section(class_name);
+
+    if (auto* part_name_sect = part_name.empty() ? nullptr : hlasm_ctx.ord_ctx.get_section(part_name))
+    {
+        if (!part_name_sect->goff || part_name_sect->goff->parent != class_name_sect)
+            add_diagnostic(diagnostic_op::error_A170_section_type_mismatch(part_rng));
+        else if (op_count != 1)
+            add_diagnostic(diagnostic_op::warn_A171_operands_ignored(stmt.operands_ref().field_range));
+        hlasm_ctx.ord_ctx.set_section(*part_name_sect);
+        return;
+    }
+
+    if (!part_name.empty() && hlasm_ctx.ord_ctx.symbol_defined(part_name))
+    {
+        add_diagnostic(diagnostic_op::error_E031("symbol", part_rng));
+        return;
+    }
+
+    if (class_name_sect)
+    {
+        hlasm_ctx.ord_ctx.set_section(*class_name_sect);
+
+        if (!class_name_sect->goff || part_name.empty() && class_name_sect->goff->partitioned)
+        {
+            add_diagnostic(diagnostic_op::error_A170_section_type_mismatch(stmt.label_ref().field_range));
+            return;
+        }
+        else if (!part_name.empty() && !class_name_sect->goff->partitioned || op_count > !part_name.empty())
+        {
+            add_diagnostic(diagnostic_op::warn_A171_operands_ignored(stmt.operands_ref().field_range));
+            return;
+        }
+    }
+    else if (hlasm_ctx.ord_ctx.symbol_defined(class_name))
+    {
+        add_diagnostic(diagnostic_op::error_E031("symbol", stmt.label_ref().field_range));
+        return;
+    }
+    else
+    {
+        auto sym_loc = hlasm_ctx.processing_stack_top().get_location();
+        sym_loc.pos.column = 0;
+        class_name_sect = hlasm_ctx.ord_ctx.create_and_set_class(
+            class_name, std::move(sym_loc), lib_info, nullptr, !part_name.empty());
+
+        // TODO: sectalign? part
+    }
+
+    if (part_name.empty())
+        return;
+
+    auto sym_loc = hlasm_ctx.processing_stack_top().get_location();
+    sym_loc.pos.column = 0;
+    hlasm_ctx.ord_ctx.create_and_set_class(part_name, std::move(sym_loc), lib_info, class_name_sect, false);
+}
+
+void asm_processor::process_CATTR(rebuilt_statement&& stmt)
+{
+    static constexpr size_t max_class_name_length = 16;
+    context::id_index class_name = find_label_symbol(stmt);
+    auto [op_count, part, part_rng] = cattr_ops(stmt.operands_ref());
+
+    if (class_name.empty() || class_name.to_string_view().size() > max_class_name_length)
+        add_diagnostic(diagnostic_op::error_A167_CATTR_label(stmt.label_ref().field_range));
+    else if (!hlasm_ctx.goff())
+        ; // NOP
+    else if (!hlasm_ctx.ord_ctx.get_last_active_control_section())
+        add_diagnostic(diagnostic_op::error_A169_no_section(stmt.stmt_range_ref()));
+    else
+        handle_cattr_ops(class_name, part, part_rng, op_count, stmt);
+
+    context::ordinary_assembly_dependency_solver dep_solver(hlasm_ctx.ord_ctx, lib_info);
+    hlasm_ctx.ord_ctx.symbol_dependencies().add_postponed_statement(
+        std::make_unique<postponed_statement_impl>(std::move(stmt), hlasm_ctx.processing_stack()),
+        std::move(dep_solver).derive_current_dependency_evaluation_context());
+}
+
+void asm_processor::process_XATTR(rebuilt_statement&& stmt)
+{
+    context::id_index class_name = find_label_symbol(stmt);
+
+    if (!hlasm_ctx.goff())
+        add_diagnostic(diagnostic_op::error_A166_GOFF_required(stmt.instruction_ref().field_range));
+    else if (class_name.empty())
+        add_diagnostic(diagnostic_op::error_A168_XATTR_label(stmt.label_ref().field_range));
+    else
+    {
+        for (const auto& op : stmt.operands_ref().value)
+        {
+            const auto* asm_op = op->access_asm();
+            if (!asm_op)
+                continue;
+            const auto* complex = asm_op->access_complex();
+            if (!complex)
+                continue;
+            if (complex->value.identifier != "PSECT" || complex->value.values.size() != 1)
+                continue;
+            const auto* str = dynamic_cast<const semantics::complex_assembler_operand::string_value_t*>(
+                complex->value.values.front().get());
+
+            auto [valid, psect] = hlasm_ctx.try_get_symbol_name(str->value);
+            if (!valid)
+                continue;
+
+            if (!hlasm_ctx.register_psect(class_name, psect))
+                add_diagnostic(diagnostic_op::warn_A172_psect_redefinition(complex->operand_range));
+
+            break;
+        }
+    }
+
+    context::ordinary_assembly_dependency_solver dep_solver(hlasm_ctx.ord_ctx, lib_info);
+    hlasm_ctx.ord_ctx.symbol_dependencies().add_postponed_statement(
+        std::make_unique<postponed_statement_impl>(std::move(stmt), hlasm_ctx.processing_stack()),
+        std::move(dep_solver).derive_current_dependency_evaluation_context());
 }
 
 } // namespace hlasm_plugin::parser_library::processing