+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## ****Unreleased****
 
+#### Added
+- Support CATTR, XATTR and R-type constants
+
 ## [1.14.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.13.0...1.14.0) (2024-07-16)
 
 #### Added